requirements.txt
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
faiss-cpu==1.13.0
filelock==3.20.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.13.0
networkx==3.6
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.1
pyarrow==21.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
